HTBHF-2504 rename steps to journey definitions
