+++ src/web/routes/application/index.js
@@ -1,5 +1,5 @@
 const { registerFormRoutes } = require('./register-form-routes')
-const { steps } = require('./steps')
+const { steps } = require('./journey-definitions')
 
 module.exports = {
   registerFormRoutes,
